探索アルゴリズムの最適化
